src/requirements_camera_vision.txt
# Requirements for Camera Integration and Vision API

# Flask and WebSocket support
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1

# Computer Vision and Image Processing
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1

# AI Vision Components
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8
transformers>=4.30.0
sentence-transformers>=2.2.2

# OCR Support
pytesseract>=0.3.10
easyocr>=1.7.0

# Face Recognition
face-recognition>=1.3.0
dlib>=19.24.2

# Image Processing
scipy>=1.11.0
scikit-image>=0.21.0
Pillow>=10.0.0
matplotlib>=3.7.0

# Network and Camera Communication
requests>=2.31.0
urllib3>=2.0.0
aiohttp>=3.8.0

# Scheduling and Automation
schedule>=1.2.0
python-crontab>=2.7.1

# Notifications (Platform-specific)
# Windows
win10toast>=0.9  # Windows only
pyttsx3>=2.90  # Cross-platform TTS

# macOS
# pync  # macOS only - install manually if needed

# Linux
# notify2  # Linux only - install manually if needed

# Email Support
secure-smtplib>=0.1.1

# Configuration and Data
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Database (for face recognition storage)
sqlite3  # Built-in with Python

# Networking and Discovery
python-nmap>=0.7.1
scapy>=2.5.0  # Optional for advanced network scanning

# Multiprocessing and Threading
concurrent-futures>=3.1.1  # Built-in with Python 3.2+

# Logging and Monitoring
colorlog>=6.7.0
structlog>=23.1.0

# Optional: Advanced features
# Home automation integration
# paho-mqtt>=1.6.0  # For MQTT integration
# homeassistant>=2023.0.0  # For Home Assistant integration

# Optional: Cloud storage for recordings
# boto3>=1.28.0  # For AWS S3
# google-cloud-storage>=2.10.0  # For Google Cloud Storage

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
